# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations/geocode:
    get:
      operationId: geocodeLocation
      tags: [locations]
      summary: Find a location by text
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: City, area, landmark, or address to search
      responses:
        "200":
          description: Geocoded location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationResult"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Geocoding service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/nearby:
    get:
      operationId: findNearbyPlaces
      tags: [locations]
      summary: Find nearby construction-related places
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: radiusMeters
          in: query
          required: false
          schema:
            type: integer
            minimum: 500
            maximum: 10000
            default: 5000
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Optional QadeerBuilds category filter
      responses:
        "200":
          description: Nearby places
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyPlacesResponse"
        "400":
          description: Invalid coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Nearby search service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /estimates:
    get:
      operationId: listConstructionEstimates
      tags: [estimates]
      summary: List the signed-in user's saved construction estimates
      responses:
        "200":
          description: Saved construction estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConstructionEstimate"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load saved estimates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createConstructionEstimate
      tags: [estimates]
      summary: Save a construction estimate for the signed-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConstructionEstimateBody"
      responses:
        "201":
          description: Saved construction estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructionEstimate"
        "400":
          description: Invalid estimate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to save estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    LocationResult:
      type: object
      properties:
        displayName:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        source:
          type: string
      required: [displayName, latitude, longitude, source]
    Place:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        distanceKm:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        osmUrl:
          type: ["string", "null"]
      required: [id, name, category, latitude, longitude, distanceKm, address, phone, website, osmUrl]
    NearbyPlacesResponse:
      type: object
      properties:
        center:
          $ref: "#/components/schemas/LocationResult"
        places:
          type: array
          items:
            $ref: "#/components/schemas/Place"
        source:
          type: string
        sourceNotice:
          type: string
      required: [center, places, source, sourceNotice]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    CreateConstructionEstimateBody:
      type: object
      properties:
        location:
          type: string
          minLength: 2
        plotSize:
          type: number
          exclusiveMinimum: 0
          maximum: 250000
        plotUnit:
          type: string
          enum: [marla, kanal, squareFeet, squareYards]
        plotAreaSqFt:
          type: integer
          minimum: 1
          maximum: 250000
        coveredAreaMode:
          type: string
          enum: [squareFeet, percentage]
        coveredAreaBasis:
          type: string
          enum: [perFloor, total]
        coveredAreaValue:
          type: number
          exclusiveMinimum: 0
          maximum: 250000
        floorCount:
          type: integer
          minimum: 1
          maximum: 10
        floors:
          type: string
          minLength: 1
        constructionType:
          type: string
          enum: [greyStructure, complete]
        quality:
          type: string
          enum: [basic, standard, premium]
        totalCoveredAreaSqFt:
          type: integer
          minimum: 1
          maximum: 250000
        greyMin:
          type: integer
          minimum: 0
        greyMax:
          type: integer
          minimum: 0
        finishingMin:
          type: integer
          minimum: 0
        finishingMax:
          type: integer
          minimum: 0
        estimatedMin:
          type: integer
          minimum: 0
        estimatedMax:
          type: integer
          minimum: 0
        costPerSqFtMin:
          type: integer
          minimum: 0
        costPerSqFtMax:
          type: integer
          minimum: 0
      required:
        - location
        - plotSize
        - plotUnit
        - plotAreaSqFt
        - coveredAreaMode
        - coveredAreaBasis
        - coveredAreaValue
        - floorCount
        - floors
        - constructionType
        - quality
        - totalCoveredAreaSqFt
        - greyMin
        - greyMax
        - finishingMin
        - finishingMax
        - estimatedMin
        - estimatedMax
        - costPerSqFtMin
        - costPerSqFtMax
    ConstructionEstimate:
      allOf:
        - $ref: "#/components/schemas/CreateConstructionEstimateBody"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
          required: [id, createdAt]
...